refactor build configuration files
    mysql tunnel for local dev
    LB healthcheck timeout changes
