+++ 05-quotes/go.mod
@@ -1,0 +1,7 @@
+module quotes
+
+require (
+	github.com/aws/aws-sdk-go v1.22.3
+	github.com/go-sql-driver/mysql v1.4.1
+	github.com/gorilla/mux v1.7.3
+)
