--- quotes/go.mod
@@ -1,6 +1,0 @@
-module quotes
-
-require (
-	github.com/aws/aws-sdk-go v1.22.3
-	github.com/gorilla/mux v1.7.3
-)
